requirements.txt
Django==5.0.6
djangorestframework==3.16.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==3.0.1
python-dotenv==1.1.1
asgiref==3.9.1
sqlparse==0.5.3
packaging==25.0
supabase>=2.0.0

